refactor plan modifier to validator
